AGPushAnalyticsTests in separate Test class
